Feat: 목표 보드 API 연동

+++ lib/model/response/get_single_board.dart
@@ -1,3 +1,5 @@
+import 'dart:developer';
+
 import 'package:moing_flutter/model/response/single_board_team_info.dart';
 
 class SingleBoardData {
@@ -8,7 +10,7 @@ class SingleBoardData {
 
   factory SingleBoardData.fromJson(Map<String, dynamic> json) {
     return SingleBoardData(
-      boardNum: json['boardNum'] ?? 0,
+      boardNum: (json['boardNum'] ?? 0) as int,
       teamInfo: TeamInfo.fromJson(json['teamInfo']),
     );
   }
